Refactor/dproxy: Add cool new features to dproxy and improve stability (#32)
* remove log fatal

* fix logging

* add dumping conn and more logging

* seems like I fixed it

* try more things

* revert ssl experiments

* revert ssl experiments in lockdownhandler.go

* remove super verbose log

* added recover call so socket is moved back in case of a panic

* accidentally removed close call

* use uuid in socket name to support multiple instances of the dproxy

* add command line for full binary mode

* add docs

+++ ios/debugproxy/socket_mover.go
@@ -2,18 +2,20 @@ package debugproxy
 
 import (
 	"fmt"
-	"log"
 	"os"
+
+	"github.com/google/uuid"
+	log "github.com/sirupsen/logrus"
 )
 
-const realSocketSuffix = ".real_socket"
+var realSocketSuffix = fmt.Sprintf(".%s.real_socket", uuid.New().String())
 
 func MoveSock(socket string) (string, error) {
 	newLocation := socket + realSocketSuffix
 	if fileExists(newLocation) {
 		return "", fmt.Errorf("there is already a file named: %s please remove it or restore original usbmuxd before starting the proxy", newLocation)
 	}
-	log.Printf("Moving socket %s to %s", socket, newLocation)
+	log.Infof("Moving socket %s to %s", socket, newLocation)
 	err := os.Rename(socket, newLocation)
 	return newLocation, err
 }
@@ -28,12 +30,19 @@ func fileExists(filename string) bool {
 
 func MoveBack(socket string) error {
 	newLocation := socket + realSocketSuffix
-	log.Printf("Deleting fake socket %s", socket)
+	log.Infof("checking if '%s' exists", newLocation)
+	if !fileExists(newLocation) {
+		log.Infof("'%s' does not exist, doing nothing", newLocation)
+		return nil
+	}
+	log.Infof("found '%s', deleting '%s'", newLocation, socket)
+
+	log.Infof("Deleting fake socket %s", socket)
 	err := os.Remove(socket)
 	if err != nil {
-		log.Printf("Warn: failed deleting %s with error %e", socket, err)
+		log.Warnf("Failed deleting %s with error %e", socket, err)
 	}
-	log.Printf("Moving back socket %s to %s", newLocation, socket)
+	log.Infof("Moving back socket %s to %s", newLocation, socket)
 	err = os.Rename(newLocation, socket)
 	return err
 }
